tuntest: split out testing package
